reorg: make make_config to its own file
