+++ .pre-commit-config.yaml
@@ -8,7 +8,7 @@ repos:
     rev: 3.9.2
     hooks:
     - id: flake8
-      args: [--ignore, "F811,D1,D205,D209,D213,D400,D401,D999,D202,E203,E501,W503,E721,F403,F405,F401",
+      args: [--ignore, "F811,D1,D205,D209,D213,D400,D401,D999,D202,E203,E501,W503,E721,F403,F405",
       --exclude, "versioneer.py,docs/*,tests/annotations/*"]
 -   repo: https://github.com/pre-commit/mirrors-isort
     rev: v5.9.3
