+++ src/hydra_zen/structured_configs/__init__.py
@@ -1,14 +1,8 @@
 # Copyright (c) 2022 Massachusetts Institute of Technology
 # SPDX-License-Identifier: MIT
 
-from ._implementations import (
-    ZenField,
-    builds,
-    hydrated_dataclass,
-    just,
-    make_config,
-    mutable_value,
-)
+from ._implementations import builds, hydrated_dataclass, just, mutable_value
+from ._make_config import ZenField, make_config
 from ._make_custom_builds import make_custom_builds_fn
 
 __all__ = [
